Add in legacy ns;
