--- ruby/lib/cucumber/query/errors.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Cucumber
-  module Query
-    TestCaseUnknownError = Class.new(StandardError)
-    TestStepUnknownError = Class.new(StandardError)
-  end
-end
