+++ ruby/lib/cucumber/query/legacy/errors.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Cucumber
+  module Query
+    module Legacy
+      TestCaseUnknownError = Class.new(StandardError)
+      TestStepUnknownError = Class.new(StandardError)
+    end
+  end
+end
