+++ ruby/lib/cucumber/query/legacy/test_run_started.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require_relative 'errors'
+
+module Cucumber
+  module Query
+    module Legacy
+      class TestRunStarted
+        def initialize(config)
+          @config = config
+        end
+
+        def id
+          @id ||= @config.id_generator.new_id
+        end
+      end
+    end
+  end
+end
